backend/requirements.txt
fastapi
uvicorn
python-dotenv
langchain-core
langgraph
langchain-google-genai
langchain-groq
tavily-python
requests
beautifulsoup4
newspaper3k
lxml[html_clean]
transformers
torch
Pillow
httpx
duckduckgo-search
multipart